requirements.txt
streamlit
pandas
plotly
sqlalchemy
prophet
python-dotenv
requests
openpyxl
